Implement dialog service for import modals
